Implemented Bezier convex hull approximation

+++ BezierConvexHull/BezierConvexHull/Model/Helpers.cs
@@ -15,17 +15,17 @@ namespace BezierConvexHull
 		/// <param name="p2">End point of the line</param>
 		/// <param name="p">Point to localize</param>
 		/// <returns>Cross product of vectors p1p2 and p1p</returns>
-		public static int CrossProduct(Point p1, Point p2, Point p)
+		public static double CrossProduct(Point p1, Point p2, Point p)
 		{
 			return (p2.x - p1.x) * (p.y - p1.y) - (p.x - p1.x) * (p2.y - p1.y);
 		}
 
-		public static int CrossProductSign(Point p1, Point p2, Point p)
+		public static double CrossProductSign(Point p1, Point p2, Point p)
 		{
 			return Math.Sign((p2.x - p1.x) * (p.y - p1.y) - (p.x - p1.x) * (p2.y - p1.y));
 		}
 
-		public static int CrossProductLength(Point p1, Point p2, Point p)
+		public static double CrossProductLength(Point p1, Point p2, Point p)
 		{
 			return Math.Abs((p2.x - p1.x) * (p.y - p1.y) - (p.x - p1.x) * (p2.y - p1.y));
 		}
@@ -37,15 +37,15 @@ namespace BezierConvexHull
 		/// <param name="q"></param>
 		/// <param name="r"></param>
 		/// <returns>0 - p, q and r are colinear, 1 - Clockwise, -1 - Counterclockwise</returns>
-		public static int Orientation(Point p, Point q, Point r)
+		public static double Orientation(Point p, Point q, Point r)
 		{
-			int value = (q.y - p.y) * (r.x - q.x) - (q.x - p.x) * (r.y - q.y);
+			double value = (q.y - p.y) * (r.x - q.x) - (q.x - p.x) * (r.y - q.y);
 
 			if (value == 0) return 0; // colinear
 			return (value > 0) ? 1 : -1;
 		}
 
-		public static int DistanceSquare(Point p1, Point p2)
+		public static double DistanceSquare(Point p1, Point p2)
 		{
 			return (p1.x - p2.x) * (p1.x - p2.x) + (p1.y - p2.y) * (p1.y - p2.y);
 		}
