+++ BezierConvexHull/BezierConvexHull/ViewModel/PointSetViewModel.cs
@@ -13,13 +13,13 @@ namespace BezierConvexHull
 
         public int PointsNumber { get; set; } = 15;
 
-        public int MinX { get; set; }
+        public double MinX { get; set; }
 
-        public int MaxX { get; set; }
+        public double MaxX { get; set; }
 
-        public int MinY { get; set; }
+        public double MinY { get; set; }
 
-        public int MaxY { get; set; }   
+        public double MaxY { get; set; }   
 
         public PointSetViewModel(List<PointViewModel> points)
         {
